use enabledChains to isolate hooks from AppProviders for testing purposes
